Move unique host functionality out of router plugin
