+++ plugins/router/template/plugin_test.go
@@ -12,6 +12,7 @@ import (
 	"k8s.io/kubernetes/pkg/watch"
 
 	routeapi "github.com/openshift/origin/pkg/route/api"
+	"github.com/openshift/origin/pkg/router/controller"
 )
 
 // TestRouter provides an implementation of the plugin's router interface suitable for unit testing.
@@ -201,7 +202,10 @@ func TestHandleEndpoints(t *testing.T) {
 	}
 
 	router := newTestRouter(make(map[string]ServiceUnit))
-	plugin := newDefaultTemplatePlugin(router)
+	templatePlugin := newDefaultTemplatePlugin(router)
+	// TODO: move tests that rely on unique hosts to pkg/router/controller and remove them from
+	// here
+	plugin := controller.NewUniqueHost(templatePlugin, controller.HostForRoute)
 
 	for _, tc := range testCases {
 		plugin.HandleEndpoints(tc.eventType, tc.endpoints)
@@ -210,7 +214,7 @@ func TestHandleEndpoints(t *testing.T) {
 			t.Errorf("Expected router to be committed after HandleEndpoints call")
 		}
 
-		su, ok := plugin.Router.FindServiceUnit(tc.expectedServiceUnit.Name)
+		su, ok := router.FindServiceUnit(tc.expectedServiceUnit.Name)
 
 		if !ok {
 			t.Errorf("TestHandleEndpoints test case %s failed.  Couldn't find expected service unit with name %s", tc.name, tc.expectedServiceUnit.Name)
@@ -229,7 +233,10 @@ func TestHandleEndpoints(t *testing.T) {
 // TestHandleRoute test route watch events
 func TestHandleRoute(t *testing.T) {
 	router := newTestRouter(make(map[string]ServiceUnit))
-	plugin := newDefaultTemplatePlugin(router)
+	templatePlugin := newDefaultTemplatePlugin(router)
+	// TODO: move tests that rely on unique hosts to pkg/router/controller and remove them from
+	// here
+	plugin := controller.NewUniqueHost(templatePlugin, controller.HostForRoute)
 
 	original := util.Time{time.Now()}
 
@@ -291,7 +298,7 @@ func TestHandleRoute(t *testing.T) {
 	if _, ok := router.FindServiceUnit("foo/TestService2"); ok {
 		t.Fatalf("unexpected second unit: %#v", router)
 	}
-	if r, ok := plugin.hostToRoute["www.example.com"]; !ok || r[0].Name != "test" {
+	if r, ok := plugin.RoutesForHost("www.example.com"); !ok || r[0].Name != "test" {
 		t.Fatalf("unexpected claimed routes: %#v", r)
 	}
 
@@ -305,7 +312,7 @@ func TestHandleRoute(t *testing.T) {
 	if _, ok := router.FindServiceUnit("foo/TestService"); !ok {
 		t.Fatalf("unexpected first unit: %#v", router)
 	}
-	if r, ok := plugin.hostToRoute["www.example.com"]; !ok || r[0].Name != "test" {
+	if r, ok := plugin.RoutesForHost("www.example.com"); !ok || r[0].Name != "test" {
 		t.Fatalf("unexpected claimed routes: %#v", r)
 	}
 
@@ -347,8 +354,8 @@ func TestHandleRoute(t *testing.T) {
 			}
 		}
 	}
-	if len(plugin.hostToRoute) != 1 {
-		t.Fatalf("did not clear claimed route: %#v", plugin.hostToRoute)
+	if plugin.HostLen() != 1 {
+		t.Fatalf("did not clear claimed route: %#v", plugin)
 	}
 
 	//delete
@@ -368,14 +375,17 @@ func TestHandleRoute(t *testing.T) {
 			t.Errorf("TestHandleRoute did not expect route key %s", router.routeKey(route))
 		}
 	}
-	if len(plugin.hostToRoute) != 0 {
-		t.Errorf("did not clear claimed route: %#v", plugin.hostToRoute)
+	if plugin.HostLen() != 0 {
+		t.Errorf("did not clear claimed route: %#v", plugin)
 	}
 }
 
 func TestNamespaceScopingFromEmpty(t *testing.T) {
 	router := newTestRouter(make(map[string]ServiceUnit))
-	plugin := newDefaultTemplatePlugin(router)
+	templatePlugin := newDefaultTemplatePlugin(router)
+	// TODO: move tests that rely on unique hosts to pkg/router/controller and remove them from
+	// here
+	plugin := controller.NewUniqueHost(templatePlugin, controller.HostForRoute)
 
 	// no namespaces allowed
 	plugin.HandleNamespaces(util.StringSet{})
@@ -394,7 +404,7 @@ func TestNamespaceScopingFromEmpty(t *testing.T) {
 	// ignores all events for namespace that doesn't match
 	for _, s := range []watch.EventType{watch.Added, watch.Modified, watch.Deleted} {
 		plugin.HandleRoute(s, route)
-		if _, ok := router.FindServiceUnit("foo/TestService"); ok || len(plugin.hostToRoute) != 0 {
+		if _, ok := router.FindServiceUnit("foo/TestService"); ok || plugin.HostLen() != 0 {
 			t.Errorf("unexpected router state %#v", router)
 		}
 	}
@@ -403,7 +413,7 @@ func TestNamespaceScopingFromEmpty(t *testing.T) {
 	plugin.HandleNamespaces(util.NewStringSet("bar"))
 	for _, s := range []watch.EventType{watch.Added, watch.Modified, watch.Deleted} {
 		plugin.HandleRoute(s, route)
-		if _, ok := router.FindServiceUnit("foo/TestService"); ok || len(plugin.hostToRoute) != 0 {
+		if _, ok := router.FindServiceUnit("foo/TestService"); ok || plugin.HostLen() != 0 {
 			t.Errorf("unexpected router state %#v", router)
 		}
 	}
@@ -411,21 +421,21 @@ func TestNamespaceScopingFromEmpty(t *testing.T) {
 	// allow foo
 	plugin.HandleNamespaces(util.NewStringSet("foo", "bar"))
 	plugin.HandleRoute(watch.Added, route)
-	if _, ok := router.FindServiceUnit("foo/TestService"); !ok || len(plugin.hostToRoute) != 1 {
+	if _, ok := router.FindServiceUnit("foo/TestService"); !ok || plugin.HostLen() != 1 {
 		t.Errorf("unexpected router state %#v", router)
 	}
 
 	// forbid foo, and make sure it's cleared
 	plugin.HandleNamespaces(util.NewStringSet("bar"))
-	if _, ok := router.FindServiceUnit("foo/TestService"); ok || len(plugin.hostToRoute) != 0 {
+	if _, ok := router.FindServiceUnit("foo/TestService"); ok || plugin.HostLen() != 0 {
 		t.Errorf("unexpected router state %#v", router)
 	}
 	plugin.HandleRoute(watch.Modified, route)
-	if _, ok := router.FindServiceUnit("foo/TestService"); ok || len(plugin.hostToRoute) != 0 {
+	if _, ok := router.FindServiceUnit("foo/TestService"); ok || plugin.HostLen() != 0 {
 		t.Errorf("unexpected router state %#v", router)
 	}
 	plugin.HandleRoute(watch.Added, route)
-	if _, ok := router.FindServiceUnit("foo/TestService"); ok || len(plugin.hostToRoute) != 0 {
+	if _, ok := router.FindServiceUnit("foo/TestService"); ok || plugin.HostLen() != 0 {
 		t.Errorf("unexpected router state %#v", router)
 	}
 }
